New Logic was added to start a new game when the current one is done.
We now have a counter, for the number of times we do get it right. When that reaches to 7 then it will invoke the startNewGame() function.
Inside that function we will also check to see what level they are currently in, so if they are in level 2 it goes back to level one, for now!
Also, if we click on the submit button and it is empty, nothing will happen. Before we used to subtract from the score but we changed that by adding and
if statment to only count if there is some text in the textfield.